+++ model/dto/overview.go
@@ -2,14 +2,14 @@ package dto
 
 type Overview struct {
 	// 攻击概览
-	AttackIncrement uint `json:"attack_increment" form:"attack_increment"`
-	AttackTotal     uint `json:"attack_total" form:"attack_total"`
+	AttackIncrement int64 `json:"attack_increment" form:"attack_increment"`
+	AttackTotal     int64 `json:"attack_total" form:"attack_total"`
 
 	// 正常类型
-	NormalIncrement uint `json:"normal_increment" form:"normal_increment"`
-	NormalTotal     uint `json:"normal_total" form:"normal_total"`
+	NormalIncrement int64 `json:"normal_increment" form:"normal_increment"`
+	NormalTotal     int64 `json:"normal_total" form:"normal_total"`
 
 	// 总数
-	Variation uint `json:"variation" form:"variation"`
-	Total     uint `json:"total" form:"total"`
+	Variation int64 `json:"variation" form:"variation"`
+	Total     int64 `json:"total" form:"total"`
 }
