+++ model/dto/statistics.go
@@ -2,5 +2,5 @@ package dto
 
 type Statistics struct {
 	AttackName []string `json:"attack_name" form:"attack_name"`
-	AttackNum  []uint   `json:"attack_num" form:"attack_num"`
+	AttackNum  []int64  `json:"attack_num" form:"attack_num"`
 }
